Simulate trait data

For a given phylogenetic network generate tip data. Any values not specified use default values

SimulateTips(phy.graph, sigma.sq = 1, mu = 0, bt = 1, vh = 0,
  SE = 0, measurement.error = 0, gamma = 0.5,
  exclude.donors.recipients = TRUE)

Arguments

phy.graph
An ape::evonet object (a phylogeny stored in phylo format that also includes a reticulation matrix)

sigma.sq
The Brownian motion wiggle rate

mu
The population mean (in the absence of hybridization)

bt
The beta value (multiplier on expected value for each hybridization event)

vh
The burst of variance that comes from a hybridization event

SE
Uniform uncertainty at the tips

measurement.error
Uncertainty at the tips, especially if it varies between species

gamma
In a hybridization event, what proportion of the trait comes from the donating parent. 0.5 means half comes from each parent

exclude.donors.recipients
If TRUE, do not generate for any donors or recipient placeholder taxa

Value

A vector of trait values

Examples

network <- SimulateNetwork(ntax=5, nhybridizations=2)
tips <- SimulateTips(network, mu=1.1, bt=3, vh=1.1, SE=1)